+++ src/WebExpress.WebIndex.Test/Token/UnitTestIndexPipeStageLowerCase.cs
@@ -45,9 +45,11 @@ namespace WebExpress.WebIndex.Test.Token
         [InlineData("Families", "families", "en")]
         public void LowerCase(string term, string normalizeTerm, string cultureString)
         {
+            // preconditions
             var culture = CultureInfo.GetCultureInfo(cultureString);
             var pipeStage = new IndexPipeStageConverterLowerCase(Fixture.Context);
 
+            // test execution
             var res = pipeStage.Process(IndexTermTokenizer.Tokenize(term, culture), culture)
                 .FirstOrDefault();
