+++ src/WebExpress.WebIndex.Test/Token/UnitTestIndexPipeStageStopWord.cs
@@ -38,6 +38,18 @@ namespace WebExpress.WebIndex.Test.Token
             "May the force be with you.",
             "may", "the", "be", "with", "you"
         )]
+        [InlineData
+        (
+            "en",
+            "Would you like tea or coffee?",
+            "would", "you", "like", "or", "coffee"
+        )]
+        [InlineData
+        (
+            "en",
+            "If it rains tomorrow, we will stay indoors.",
+            "if", "it", "we", "will", "stay"
+        )]
         [Theory]
         [InlineData
         (
@@ -53,17 +65,18 @@ namespace WebExpress.WebIndex.Test.Token
         )]
         public void StopWord(string cultureStr, string str, params string[] tokenStr)
         {
+            // preconditions
             var culture = CultureInfo.GetCultureInfo(cultureStr);
             var pipeStage = new IndexPipeStageFilterStopWord(Fixture.Context);
 
             var token = IndexTermTokenizer.Tokenize(str.ToLower(), culture);
 
+            // test execution
             var res = pipeStage.Process(token, culture)
                 .Select(x => x.Value)
                 .ToList();
 
             Assert.DoesNotContain(tokenStr, res);
-            Assert.True(token.Count() - (tokenStr?.Length ?? 0) == res.Count);
         }
     }
 }
