+++ src/WebExpress.WebIndex/Memory/IndexMemorySchema.cs
@@ -1,4 +1,7 @@
-﻿namespace WebExpress.WebIndex.Memory
+﻿using System;
+using System.Collections.Generic;
+
+namespace WebExpress.WebIndex.Memory
 {
     /// <summary>
     /// Represents a index schema file.
@@ -11,6 +14,11 @@
         /// Returns the index context.
         /// </summary>
         public IIndexContext Context { get; private set; }
+
+        /// <summary>
+        /// Return the index field data.
+        /// </summary>
+        public IEnumerable<IndexFieldData> Fields => GetFieldData(typeof(TIndexItem));
 
         /// <summary>
         /// Initializes a new instance of the <see cref="IndexSchema"/> class.
@@ -52,7 +60,46 @@
         /// </summary>
         public virtual void Dispose()
         {
+            GC.SuppressFinalize(this);
+        }
 
+        /// <summary>
+        /// Recursively retrieves the field names of the specified type.
+        /// </summary>
+        /// <param name="type">The type whose field names to retrieve.</param>
+        /// <param name="prefix">The prefix to prepend to each field name.</param>
+        /// <param name="processedTypes">A set of types that have already been processed to avoid circular references.</param>
+        /// <returns>An enumerable collection of field names.</returns>
+        private static IEnumerable<IndexFieldData> GetFieldData(Type type, string prefix = "", HashSet<Type> processedTypes = null)
+        {
+            processedTypes ??= [];
+
+            if (processedTypes.Contains(type))
+            {
+                yield break;
+            }
+
+            processedTypes.Add(type);
+
+            foreach (var property in type.GetProperties())
+            {
+                string propertyName = string.IsNullOrEmpty(prefix) ? property.Name : $"{prefix}.{property.Name}";
+
+                yield return new IndexFieldData
+                {
+                    Name = propertyName,
+                    Type = property.PropertyType,
+                    PropertyInfo = property
+                };
+
+                if (property.PropertyType.IsClass && property.PropertyType != typeof(string))
+                {
+                    foreach (var subProperty in GetFieldData(property.PropertyType, propertyName, processedTypes))
+                    {
+                        yield return subProperty;
+                    }
+                }
+            }
         }
     }
 }
