+++ src/WebExpress.WebIndex/Storage/IndexStorageSchema.cs
@@ -1,9 +1,9 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Reflection;
 using System.Text.Json;
-using WebExpress.WebIndex.WebAttribute;
 
 namespace WebExpress.WebIndex.Storage
 {
@@ -28,6 +28,11 @@ namespace WebExpress.WebIndex.Storage
         /// Returns the index context.
         /// </summary>
         public IIndexContext Context { get; private set; }
+
+        /// <summary>
+        /// Return the index field data.
+        /// </summary>
+        public IEnumerable<IndexFieldData> Fields => GetFieldData(typeof(TIndexItem));
 
         /// <summary>
         /// Initializes a new instance of the class.
@@ -89,12 +94,12 @@ namespace WebExpress.WebIndex.Storage
         private dynamic GetSchema()
         {
             var objectType = typeof(TIndexItem);
-            var fields = objectType.GetProperties().Select(x => new
+            var fields = Fields.Select(x => new
             {
                 x.Name,
-                Type = GetType(x),
-                Ignore = Attribute.IsDefined(x, typeof(IndexIgnoreAttribute)),
-                Abstract = x.GetGetMethod().IsAbstract
+                Type = GetType(x.PropertyInfo),
+                Ignore = !x.Enabled,
+                Abstract = x.PropertyInfo.GetGetMethod().IsAbstract
             });
             var schema = new { objectType.Name, Fields = fields };
 
@@ -169,6 +174,45 @@ namespace WebExpress.WebIndex.Storage
         public virtual void Dispose()
         {
             GC.SuppressFinalize(this);
+        }
+
+        /// <summary>
+        /// Recursively retrieves the field names of the specified type.
+        /// </summary>
+        /// <param name="type">The type whose field names to retrieve.</param>
+        /// <param name="prefix">The prefix to prepend to each field name.</param>
+        /// <param name="processedTypes">A set of types that have already been processed to avoid circular references.</param>
+        /// <returns>An enumerable collection of field names.</returns>
+        private static IEnumerable<IndexFieldData> GetFieldData(Type type, string prefix = "", HashSet<Type> processedTypes = null)
+        {
+            processedTypes ??= [];
+
+            if (processedTypes.Contains(type))
+            {
+                yield break;
+            }
+
+            processedTypes.Add(type);
+
+            foreach (var property in type.GetProperties())
+            {
+                string propertyName = string.IsNullOrEmpty(prefix) ? property.Name : $"{prefix}.{property.Name}";
+
+                yield return new IndexFieldData
+                {
+                    Name = propertyName,
+                    Type = property.PropertyType,
+                    PropertyInfo = property
+                };
+
+                if (property.PropertyType.IsClass && property.PropertyType != typeof(string))
+                {
+                    foreach (var subProperty in GetFieldData(property.PropertyType, propertyName, processedTypes))
+                    {
+                        yield return subProperty;
+                    }
+                }
+            }
         }
     }
 }
